refactor: hydrate 적용 방식 변경 (dept)
