+++ apps/community/src/components/about/dept/dept-info-section.tsx
@@ -1,0 +1,35 @@
+'use client';
+
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { Fragment } from 'react';
+
+import { sprinkles } from '@aics-client/design-system/styles';
+
+import { deptQueryOptions } from '~/apis/about/dept/queries';
+
+import { List } from '~/components/about/list';
+import { Section } from '~/components/about/section';
+
+function DeptInfoSection() {
+  const { data } = useSuspenseQuery(deptQueryOptions.all());
+
+  return (
+    <Section>
+      {data.data.map((dept) => (
+        <Fragment key={`dept-${dept.name}`}>
+          <Section.Title>{dept.name}</Section.Title>
+          <p className={sprinkles({ marginBottom: 'none' })}>
+            {dept.description}
+          </p>
+          <List title="교육 목표">
+            {dept.educationGoals.map((goal) => (
+              <List.Row key={`dept-goal-${goal}`}>{goal}</List.Row>
+            ))}
+          </List>
+        </Fragment>
+      ))}
+    </Section>
+  );
+}
+
+export { DeptInfoSection };
