--- apps/community/src/types/dept.ts
@@ -1,7 +1,0 @@
-interface Dept {
-  name: string;
-  description: string;
-  educationGoals: string[];
-}
-
-export type { Dept };
